adjust the test to by sync test
