--- tests/dummy.test.ts
@@ -1,5 +1,0 @@
-import { test, expect } from 'vitest';
-
-test('dummy', () => {
-  expect(1).toBe(1);
-});
